Give every type its own file and let the scanner read its own config
The package had several unrelated types sharing files, and the scanner was
built by pulling values out of config in the container, which also meant
the container decided what an unset url means.

Each struct now lives in a file named after it, the scanner takes the
config interface and checks it in Start like the certificate loader does,
and the loader is named for what it does rather than what it holds.
