+++ backend/mailrelay/counters.go
@@ -1,0 +1,7 @@
+package mailrelay
+
+type counters struct {
+	minute window
+	hour   window
+	day    window
+}
