+++ backend/mailrelay/limits.go
@@ -1,0 +1,8 @@
+package mailrelay
+
+type Limits struct {
+	Minute     int64
+	Hour       int64
+	Day        int64
+	Recipients int
+}
