+++ backend/mailrelay/rspamd_result.go
@@ -1,0 +1,6 @@
+package mailrelay
+
+type rspamdResult struct {
+	Action string  `json:"action"`
+	Score  float64 `json:"score"`
+}
