+++ backend/mailrelay/sns_notification.go
@@ -1,0 +1,6 @@
+package mailrelay
+
+type snsNotification struct {
+	Type    string `json:"Type"`
+	Message string `json:"Message"`
+}
