+++ backend/mailrelay/window.go
@@ -1,0 +1,8 @@
+package mailrelay
+
+import "time"
+
+type window struct {
+	start time.Time
+	count int64
+}
